feat: added category modules for all domain configuration
